# docs/api/swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Skyline API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/setting/{key}": {
            "get": {
                "tags": [
                    "Setting"
                ],
                "summary": "Show Setting",
                "description": "Get a setting item.",
                "operationId": "show_setting_api_v1_setting__key__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Key",
                            "type": "string"
                        },
                        "name": "key",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Setting"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedMessage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundMessage"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Setting"
                ],
                "summary": "Reset Setting",
                "description": "Reset a setting item to default",
                "operationId": "reset_setting_api_v1_setting__key__delete",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Key",
                            "type": "string"
                        },
                        "name": "key",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedMessage"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenMessage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundMessage"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/setting": {
            "put": {
                "tags": [
                    "Setting"
                ],
                "summary": "Update Setting",
                "description": "Update a setting item.",
                "operationId": "update_setting_api_v1_setting_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSetting"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Setting"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedMessage"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenMessage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundMessage"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityMessage"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/settings": {
            "get": {
                "tags": [
                    "Setting"
                ],
                "summary": "List Settings",
                "description": "Get all settings.",
                "operationId": "list_settings_api_v1_settings_get",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Settings"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedMessage"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ForbiddenMessage": {
                "title": "ForbiddenMessage",
                "required": [
                    "detail"
                ],
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "string"
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "NotFoundMessage": {
                "title": "NotFoundMessage",
                "required": [
                    "detail"
                ],
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "string"
                    }
                }
            },
...
            "Setting": {
                "title": "Setting",
                "required": [
                    "key",
                    "restart_service"
                ],
                "type": "object",
                "properties": {
                    "key": {
                        "title": "Key",
                        "type": "string"
                    },
                    "value": {
                        "title": "Value"
                    },
                    "hidden": {
                        "title": "Hidden",
                        "type": "boolean"
                    },
                    "restart_service": {
                        "title": "Restart Service",
                        "type": "boolean"
                    }
                }
            },
            "Settings": {
                "title": "Settings",
                "required": [
                    "settings"
                ],
                "type": "object",
                "properties": {
                    "settings": {
                        "title": "Settings",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Setting"
                        }
                    }
                }
            },
            "UnauthorizedMessage": {
                "title": "UnauthorizedMessage",
                "required": [
                    "detail"
                ],
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "string"
                    }
                }
            },
            "UnprocessableEntityMessage": {
                "title": "UnprocessableEntityMessage",
                "required": [
                    "detail"
                ],
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "string"
                    }
                }
            },
            "UpdateSetting": {
                "title": "UpdateSetting",
                "required": [
                    "key"
                ],
                "type": "object",
                "properties": {
                    "key": {
                        "title": "Key",
                        "type": "string"
                    },
                    "value": {
                        "title": "Value"
                    }
                }
            },
...
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            },
...
        }
    }
}